+++ packages/react-resizable-panels/src/types.ts
@@ -7,3 +7,7 @@ export type MixedSizes = {
 
 export type ResizeEvent = KeyboardEvent | MouseEvent | TouchEvent;
 export type ResizeHandler = (event: ResizeEvent) => void;
+
+export type DataAttributes = {
+  [attribute: string]: string | number | boolean | undefined;
+};